Inline the schema and remove their tests
We will add coverage wholisticly later
