+++ lms/validation/authentication/__init__.py
@@ -11,25 +11,4 @@ from lms.validation.authentication._exceptions import (
     MissingStateParamError,
 )
 from lms.validation.authentication._launch_params import LaunchParamsAuthSchema
-from lms.validation.authentication._oauth import (
-    CanvasAccessTokenResponseSchema,
-    CanvasOAuthCallbackSchema,
-    CanvasRefreshTokenResponseSchema,
-)
-
-__all__ = (
-    "LaunchParamsAuthSchema",
-    "CanvasAccessTokenResponseSchema",
-    "CanvasOAuthCallbackSchema",
-    "CanvasRefreshTokenResponseSchema",
-    "BearerTokenSchema",
-    "ExpiredSessionTokenError",
-    "MissingSessionTokenError",
-    "InvalidSessionTokenError",
-    "MissingStateParamError",
-    "InvalidStateParamError",
-    "ExpiredStateParamError",
-    "JWTError",
-    "ExpiredJWTError",
-    "InvalidJWTError",
-)
+from lms.validation.authentication._oauth import CanvasOAuthCallbackSchema
